Move some code out of blockchain into heuristics
