+++ src/python-interface/python_interface.cpp
@@ -8,6 +8,7 @@
 
 #include "optional_py.hpp"
 #include "variant_py.hpp"
+#include "input_py.hpp"
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
@@ -18,7 +19,6 @@ namespace py = pybind11;
 void init_address(py::module &m);
 void init_tx(py::module &m);
 void init_tx_summary(py::module &m);
-void init_input(py::module &m);
 void init_output(py::module &m);
 void init_block(py::module &m);
 void init_blockchain(py::module &m);
